handle json format 64bit integer issue #334

+++ dots_server/db_models/telemetry_source_count.go
@@ -6,10 +6,10 @@ import "github.com/go-xorm/xorm"
 type TelemetrySourceCount struct {
 	Id                 int64     `xorm:"'id' pk autoincr"`
 	TeleAttackDetailId int64     `xorm:"'tele_attack_detail_id' not null"`
-	LowPercentileG     int       `xorm:"low_percentile_g"`
-	MidPercentileG     int       `xorm:"mid_percentile_g"`
-	HighPercentileG    int       `xorm:"high_percentile_g"`
-	PeakG              int       `xorm:"peak_g"`
+	LowPercentileG     uint64    `xorm:"low_percentile_g"`
+	MidPercentileG     uint64    `xorm:"mid_percentile_g"`
+	HighPercentileG    uint64    `xorm:"high_percentile_g"`
+	PeakG              uint64    `xorm:"peak_g"`
 	Created            time.Time `xorm:"created"`
 	Updated            time.Time `xorm:"updated"`
 }
